Create Email_Failed_Task.ps1
This PowerShell script is designed to check the status of a specified scheduled task. It queries the task status and checks if it has failed (i.e., if the error code is greater than 0). If the task fails, it sends an email notification to an administrator with the error code.

To ensure this check is performed daily, the script must be added to Windows Task Scheduler. The Task Scheduler will run the script at your desired frequency (e.g., daily) to check for task failures and send alerts if needed.