Create Export_All_Distribution_Groups.ps1
This script queries Active Directory to retrieve all groups where the GroupCategory is 'Distribution' 
and the GroupScope is not 'DomainLocal'. The results are exported to a CSV file in C:\Temp for 
reporting or auditing purposes. Ensure the ActiveDirectory module is installed and you have sufficient 
permissions to query AD groups.