Create Get_Logged_On_Users_Mapped_Drives.ps1
This script queries the Windows registry under HKEY_USERS to find network drives mapped for 
currently logged-on users. For each mapped drive, it outputs the username, drive letter, remote path, 
the username used to connect (if different credentials were used), and the user's SID. This information 
is returned as PSCustomObjects, allowing for further processing or export. Verbose output can indicate 
if no mapped drives are found. Note that only users with loaded profiles in HKEY_USERS are included.